+++ docs/news.rst
@@ -3,7 +3,7 @@ Release Notes
 
 **2.0.0 (unreleased)**
 
-* **BREAKING CHANGE:** Added support for plugins.
+* Added support for plugins.
 
   * This introduces plugin hooks ``pypom_after_wait_for_page_to_load`` and
     ``pypom_after_wait_for_region_to_load``.
@@ -16,6 +16,9 @@ Release Notes
     the page or region, which returns ``True`` when the page or region has
     finished loading.
   * See the user guide for more details.
+
+* Any unused ``url_kwargs`` after formatting ``URL_TEMPLATE`` are added as URL
+  query string parameters.
 
 **1.3.0 (2018-02-28)**